refactor(lib): extract shared widgets
