+++ lib/widgets.dart
@@ -2,5 +2,6 @@ library widgets;
 
 export 'src/widgets/query_wizard.dart';
 export 'src/widgets/query_wizard_view.dart';
+export 'src/widgets/shared/tree_viewer.dart';
 export 'src/widgets/tabs/tables_and_fields.dart';
 export 'src/widgets/tabs/parts/sources_tree.dart';
